captura/requirements.txt
fastapi
uvicorn
sqlalchemy>=1.4,<2.0
psycopg2-binary
python-dotenv
pyproj 
requests
rasterio
matplotlib
shapely
